protocol generator: use protobuf double type to represent float values

+++ aea/protocols/generator/common.py
@@ -69,7 +69,7 @@ SERIALIZATION_DOT_PY_FILE_NAME = "serialization.py"
 PYTHON_TYPE_TO_PROTO_TYPE = {
     "bytes": "bytes",
     "int": "int32",
-    "float": "float",
+    "float": "double",
     "bool": "bool",
     "str": "string",
 }
